prototype/requirements.txt
pandas
numpy
pymavlink
PyYAML
matplotlib
